requirements-test.txt
# Testing dependencies
pytest==7.4.0
pytest-django==4.5.2
selenium==4.11.2
webdriver-manager==3.8.6
pytest-xdist==3.3.1
pytest-cov==4.1.0
